requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.1
aiohttp>=3.8.4
aiohttp-jinja2>=1.6
aiosqlite>=0.18.0
async-timeout>=4.0.2
pyyaml>=6.0

# Optional / only if you enable music playback:
PyNaCl>=1.5.0
yt-dlp>=2025.10.01

# Optional image / meme generation (if you plug in an API)
Pillow>=10.0.0
requests>=2.31.0
